Add hex module to api
