+++ src/sys/net/mod.rs
@@ -5,7 +5,7 @@ pub mod mac;
 pub mod stat;
 pub mod socket;
 
-use crate::{sys, usr};
+use crate::{api, sys};
 use crate::sys::pci::DeviceConfig;
 
 use alloc::format;
@@ -111,7 +111,7 @@ impl<'a> smoltcp::phy::Device for EthernetDevice {
         if let Some(buffer) = self.receive_packet() {
             if self.config().is_debug_enabled() {
                 debug!("NET Packet Received");
-                usr::hex::print_hex(&buffer);
+                printk!("{}", api::hex::format_hex(&buffer));
             }
             self.stats().rx_add(buffer.len() as u64);
             let rx = RxToken { buffer };
@@ -163,7 +163,7 @@ impl smoltcp::phy::TxToken for TxToken {
         let res = f(buf);
         if config.is_debug_enabled() {
             debug!("NET Packet Transmitted");
-            usr::hex::print_hex(buf);
+            printk!("{}", api::hex::format_hex(buf));
         }
         self.device.transmit_packet(len);
         self.device.stats().tx_add(len as u64);
