+++ src/usr/host.rs
@@ -166,7 +166,7 @@ pub fn resolve(name: &str) -> Result<IpAddress, ResponseCode> {
                 let message = Message::from(&data);
                 if message.id() == query.id() && message.is_response() {
                     syscall::close(handle);
-                    //usr::hex::print_hex(&message.datagram);
+                    //print!("{}", api::hex::format_hex(&message.datagram));
                     return match message.code() {
                         ResponseCode::NoError => {
                             // TODO: Parse the datagram instead of extracting
